requirements.txt
beautifulsoup4==4.5.1
defusedxml==0.4.1
Django==1.10.2
django-appconf==1.0.2
django-compressor==2.1
django-cors-headers==1.2.2
django-filter==0.15.2
django-htmlmin==0.9.1
django-mailgun==0.9.1
django-modeltranslation==0.12
djangorestframework==3.4.7
gunicorn==19.6.0
html5lib==0.999999999
Markdown==2.6.7
oauthlib==2.0.0
Pillow==3.4.1
psycopg2==2.6.2
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
rcssmin==1.0.6
requests==2.11.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
webencodings==0.5
